Address OSM ingestion PR review feedback
Apply the outstanding PR review requirements across OSM ingestion code, tests, migration checks, and docs.

- stream CLI sha256 hashing and keep module docs ordered
- batch POI upserts and preserve non-OSM-owned columns on conflict
- make provenance persistence atomic with POI writes
- expand schema and request-validation coverage (including bounds errors)
- strengthen BDD persisted-content assertions and split scenarios
- update docs/execplan/runbook and CI lockfile practices

+++ backend/migrations/2026-02-24-000000_create_osm_ingestion_provenance/up.sql
@@ -10,8 +10,8 @@ CREATE TABLE IF NOT EXISTS osm_ingestion_provenance (
     bounds_min_lat DOUBLE PRECISION NOT NULL,
     bounds_max_lng DOUBLE PRECISION NOT NULL,
     bounds_max_lat DOUBLE PRECISION NOT NULL,
-    raw_poi_count INTEGER NOT NULL CHECK (raw_poi_count >= 0),
-    filtered_poi_count INTEGER NOT NULL CHECK (filtered_poi_count >= 0),
+    raw_poi_count BIGINT NOT NULL CHECK (raw_poi_count >= 0),
+    filtered_poi_count BIGINT NOT NULL CHECK (filtered_poi_count >= 0),
     created_at TIMESTAMPTZ NOT NULL DEFAULT NOW(),
     CONSTRAINT osm_ingestion_provenance_bounds_valid CHECK (
         bounds_min_lng >= -180 AND bounds_min_lng <= 180
